% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example survival dataset for TEP}
\format{
A data frame with the following columns:
\describe{
\item{age}{Time to event or censoring}
\item{dead}{Event indicator (1 = event, 0 = censored)}
\item{group}{Group indicator}
}
}
\usage{
example_data
}
\description{
A survival dataset included with TEP and used in package examples.
}
\examples{
head(example_data)
}
\keyword{datasets}
