% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_bs.R
\name{safe_bshazard}
\alias{safe_bshazard}
\title{Safely fit a bshazard model}
\usage{
safe_bshazard(formula, data, ..., reps = 10, age_sd = NULL)
}
\arguments{
\item{formula}{A survival model formula}

\item{data}{A data frame containing survival data}

\item{...}{Additional arguments passed to bshazard()}

\item{reps}{Number of replicated rows for data augmentation}

\item{age_sd}{Standard deviation for age jittering}
}
\value{
A fitted bshazard model object
}
\description{
This function wraps \code{bshazard::bshazard()} and automatically retries
with jittered data augmentation if a singular system error occurs.
}
